third_party/blink/web_tests/external/wpt/css/css-box/margin-trim/computed-margin-values/block-container-block-start-child-with-border-expected.txt

This is a testharness.js-based test.
FAIL item 1 assert_equals: 
<item data-expected-margin-top="0" class="collapsed">
            <item data-expected-margin-top="0" class="collapsed"></item>
        </item>
margin-top expected "0" but got "50"
FAIL item 2 assert_equals: 
<item data-expected-margin-top="0" class="collapsed"></item>
margin-top expected "0" but got "50"
FAIL item 3 assert_equals: 
<item class="with-border-and-margin" data-expected-margin-top="0">
            <item class="with-border-and-margin" data-expected-margin-top="30"></item>
        </item>
margin-top expected "0" but got "30"
FAIL item 4 assert_equals: 
<item class="with-border-and-margin" data-expected-margin-top="30"></item>
margin-top expected "0" but got "30"
Harness: the test ran to completion.
